+++ R/Class-SimParam.R
@@ -24,17 +24,44 @@ SimParam = R6Class(
     founderPop = "MapPop",
 
     #' @field finalizePop function applied to newly created populations.
+    #' It is run as the last step of creating a new population and
+    #' provides a way to automatically modify populations.
     #' The function must satisfy these four requirements:
-    #' 1) The first argument must be a \code{\link{Pop-class}} object,
-    #' 2) The second argument must be a SimParam object,
-    #' 3) The remaining arguments must be passed via \code{...}, and
-    #' 4) The return must be a \code{\link{Pop-class}} object.
+    #' 1) The first argument is a \code{\link{Pop-class}} object
+    #'    which will be used in this function;
+    #' 2) The second argument is a SimParam object
+    #'    to provide access to global parameters;
+    #' 3) The remaining arguments are passed via \code{...}; and
+    #' 4) The return is a \code{\link{Pop-class}} object.
     #' See \code{\link{asCategorical}} for an example.
     finalizePop = "function",
 
     #' @field allowEmptyPop if true, population arguments with nInd=0 will
     #' return an empty population with a warning instead of an error.
     allowEmptyPop = "logical",
+
+    #' @field finalizePheno function applied to newly generated phenotype values.
+    #' It is run as the last step of generating new phenotype values and
+    #' provides a way to automatically modify phenotype values.
+    #' While \code{finalizePop} function can be used to a similar effect,
+    #' \code{finalizePheno} function provides more granular support over
+    #' finalizing phenotype values in new and existing populations, say,
+    #' when we repeatedly call \code{\link{setPheno}} on the existing population
+    #' (which would not trigger use of \code{finalizePop}).
+    #' The function must satisfy these five requirements:
+    #' 1) The first argument is a matrix of phenotype values,
+    #'    which will be used in this function and
+    #'    should have dimensions as \code{pop@pheno};
+    #' 2) The second argument is a \code{\link{Pop-class}} object
+    #'    to provide access to population specific values;
+    #' 3) The third argument must be a SimParam object
+    #'    to provide access to global parameters;
+    #' 4) The remaining arguments must be passed via \code{...}; and
+    #' 5) The return is a matrix of phenotype values that will form the new
+    #'    \code{pop@pheno}, so the function should not change column or
+    #'    row order.
+    #' See \code{\link{asCategorical}} for an example.
+    finalizePheno = "function",
 
     #' @description Starts the process of building a new simulation
     #' by creating a new SimParam object and assigning a founder
@@ -67,6 +94,7 @@ SimParam = R6Class(
       self$founderPop = founderPop
       self$finalizePop = function(pop, ...){return(pop)}
       self$allowEmptyPop = FALSE # Empty populations trigger an error
+      self$finalizePheno = function(pheno, ...){return(pheno)}
 
       # Private items
       private$.restrSites = TRUE